requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
nltk
spacy
pdfminer.six
gradio
numpy
pandas
matplotlib
seaborn
plotly
Pillow
pyyaml
tqdm
python-dateutil
requests
scikit-learn